fix: resolve mock issues
