+++ scripts/tests/localsetup_tests.py
@@ -6,16 +6,18 @@ import unittest
 import collections
 import yaml
 
-from .. import compose
+from ..modules import cli
+from ..modules import helpers
+from ..modules import service
 from ..modules.aux_services import Postgres, Redis
 from ..modules.elastic_stack import ApmServer, Elasticsearch
-from ..modules.helpers import parse_version, resolve_bc
+from ..modules.helpers import parse_version
 from ..modules.opbeans import (
     OpbeansService, OpbeansDotnet, OpbeansGo, OpbeansJava, OpbeansNode, OpbeansPython,
     OpbeansRuby, OpbeansRum, OpbeansLoadGenerator
 )
 
-from ..compose import LocalSetup, discover_services
+from ..modules.cli import discover_services, LocalSetup
 
 from .service_tests import ServiceTest
 
@@ -875,7 +877,7 @@ class LocalTest(unittest.TestCase):
         """)  # noqa: 501
         self.assertDictEqual(got, want)
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_6_x_xpack_secure(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'6.6': '6.6.10'}):
@@ -906,7 +908,7 @@ class LocalTest(unittest.TestCase):
         ## allow anonymous healthcheck
         self.assertIn("STATUS_ALLOWANONYMOUS", kibana_env)
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_7_0_xpack_secure(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master': '8.0.0'}):
@@ -936,7 +938,7 @@ class LocalTest(unittest.TestCase):
         ## allow anonymous healthcheck
         self.assertIn("STATUS_ALLOWANONYMOUS", kibana_env)
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_no_elasticesarch(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master':'8.0.0'}):
@@ -949,7 +951,7 @@ class LocalTest(unittest.TestCase):
         self.assertNotIn("elasticsearch", services)
         self.assertNotIn("elasticsearch", services["apm-server"]["depends_on"])
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_all(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master': '8.0.0'}):
@@ -973,7 +975,7 @@ class LocalTest(unittest.TestCase):
             "postgres", "redis",
         })
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_one_opbeans(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master': '8.0.0'}):
@@ -986,7 +988,7 @@ class LocalTest(unittest.TestCase):
         self.assertIn("redis", services)
         self.assertIn("postgres", services)
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_opbeans_2nd(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master': '8.0.0'}):
@@ -1005,7 +1007,7 @@ class LocalTest(unittest.TestCase):
         self.assertIn("opbeans-python01", services)
         self.assertIn("opbeans-ruby01", services)
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_all_opbeans_no_apm_server(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         with mock.patch.dict(LocalSetup.SUPPORTED_VERSIONS, {'master': '8.0.0'}):
@@ -1023,7 +1025,7 @@ class LocalTest(unittest.TestCase):
         for name, service in got["services"].items():
             self.assertNotIn("apm-server", service.get("depends_on", {}), "{} depends on apm-server".format(name))
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_unsupported_version_pre_6_3(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         version = "1.2.3"
@@ -1040,7 +1042,7 @@ class LocalTest(unittest.TestCase):
         )
         self.assertEqual("docker.elastic.co/kibana/kibana-x-pack:{}".format(version), services["kibana"]["image"])
 
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_unsupported_version(self, _ignore_load_images):
         docker_compose_yml = stringIO()
         version = "6.9.5"
@@ -1057,8 +1059,8 @@ class LocalTest(unittest.TestCase):
         )
         self.assertEqual("docker.elastic.co/kibana/kibana:{}-SNAPSHOT".format(version), services["kibana"]["image"])
 
-    @mock.patch(compose.__name__ + ".resolve_bc")
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(service.__name__ + ".resolve_bc")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_bc(self, mock_load_images, mock_resolve_bc):
         mock_resolve_bc.return_value = {
             "projects": {
@@ -1131,8 +1133,8 @@ class LocalTest(unittest.TestCase):
             },
             image_cache_dir)
 
-    @mock.patch(compose.__name__ + ".resolve_bc")
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(service.__name__ + ".resolve_bc")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_bc_oss(self, mock_load_images, mock_resolve_bc):
         mock_resolve_bc.return_value = {
             "projects": {
@@ -1187,8 +1189,8 @@ class LocalTest(unittest.TestCase):
             },
             image_cache_dir)
 
-    @mock.patch(compose.__name__ + ".resolve_bc")
-    @mock.patch(compose.__name__ + ".load_images")
+    @mock.patch(service.__name__ + ".resolve_bc")
+    @mock.patch(cli.__name__ + ".load_images")
     def test_start_bc_with_release(self, mock_load_images, mock_resolve_bc):
         mock_resolve_bc.return_value = {
             "projects": {
@@ -1243,7 +1245,7 @@ class LocalTest(unittest.TestCase):
             },
             image_cache_dir)
 
-    @mock.patch(compose.__name__ + ".resolve_bc")
+    @mock.patch(service.__name__ + ".resolve_bc")
     def test_docker_download_image_url(self, mock_resolve_bc):
         mock_resolve_bc.return_value = {
             "projects": {
